Change class name from database_raw_access to database_read_acces
Until now the only possible interaction with an alucell database was
through read access. But I plan to implement alucell database creation
capabilities. The previous class name didn't make clear if the access
was read or write, and with both capabilities available, it is
important to make it clear.

In the futur we will have two classes to interact with an alucell
database:
 - database_read_access (which already exists),
 - database_write_access (which is to be defined).