Add an action to show the content of the block info of the dbfile
The blockinfo is a structure which hold some info necessary to
interpret the content of the dbfile, including the file offsets
where the name index, variable lengths and variable positions are
stored, as well as the file offsets of the next possible insertion.

Providing access to the content to this structure is useful typically
for debugging and troublshooting purposes.